New changes to reflect when a ship is hit or sunk
Added a new endpoint to delete game
