+++ src/main/java/com/dhaval/battleship/exceptions/BattleGridException.java
@@ -1,16 +1,16 @@
-package com.dhaval.battleship.Exceptions;
+package com.dhaval.battleship.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(value=HttpStatus.NOT_FOUND,reason = "GameId does not exist")
-public class GameDoesNotExistException extends Exception {
+@ResponseStatus(value=HttpStatus.NOT_FOUND,reason = "No Battlegrid found for supplied parameters")
+public class BattleGridException extends Exception {
     /**
      * Constructs a new exception with {@code null} as its detail message.
      * The cause is not initialized, and may subsequently be initialized by a
      * call to {@link #initCause}.
      */
-    public GameDoesNotExistException() {
+    public BattleGridException() {
     }
 
     /**
@@ -21,7 +21,7 @@ public class GameDoesNotExistException extends Exception {
      * @param message the detail message. The detail message is saved for
      *                later retrieval by the {@link #getMessage()} method.
      */
-    public GameDoesNotExistException(String message) {
+    public BattleGridException(String message) {
         super(message);
     }
 }
